split out enemy into module
